rename redux folder to fileTypes, update index.js require statements
