+++ fileTypes/component.js
