+++ fileTypes/reducer.js
