+++ fileTypes/store.js
