+++ fileTypes/util.js
